ref(onboarding): move SCM billing-config lookup into a gsApp hook
The OSS useScmFeatureMeta was reaching into getsentry behind a
boundaries/dependencies disable to read the subscription store and
billing-config endpoint. Move that work to a gsApp implementation
registered as react-hook:use-scm-feature-meta and have the OSS hook
delegate through HookStore, returning FALLBACK_FEATURE_META when no
implementation is registered.

The billing-config-aware tests move to a colocated gsApp spec; the OSS
spec keeps the rendering checks and a single fallback-path assertion.
