+++ static/gsApp/registerHooks.tsx
@@ -73,6 +73,7 @@ import {useDashboardDatasetRetentionLimit} from 'getsentry/hooks/useDashboardDat
 import {useExperiment} from 'getsentry/hooks/useExperiment';
 import {useMetricDetectorLimit} from 'getsentry/hooks/useMetricDetectorLimit';
 import {useProductBillingAccess} from 'getsentry/hooks/useProductBillingAccess';
+import {useScmFeatureMeta} from 'getsentry/hooks/useScmFeatureMeta';
 import {rawTrackAnalyticsEvent} from 'getsentry/utils/rawTrackAnalyticsEvent';
 import {trackMetric} from 'getsentry/utils/trackMetric';
 
@@ -259,6 +260,7 @@ const GETSENTRY_HOOKS: Partial<Hooks> = {
   'react-hook:use-dashboard-dataset-retention-limit': useDashboardDatasetRetentionLimit,
   'react-hook:use-experiment': useExperiment,
   'react-hook:use-product-billing-access': useProductBillingAccess,
+  'react-hook:use-scm-feature-meta': useScmFeatureMeta,
   'component:partnership-agreement': p => (
     <LazyLoad LazyComponent={PartnershipAgreement} {...p} />
   ),
